refactor(PlayStatus): Card 모듈 -> domain 모듈로 이동, BUST_CONDITION public -> default로 변경
