+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -14,6 +14,8 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import blackjack.domain.PlayStatus;
+
 class CardsTest {
 
     @ParameterizedTest
