+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -10,9 +10,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import blackjack.domain.NoShuffleDeck;
+import blackjack.domain.PlayStatus;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
-import blackjack.domain.card.PlayStatus;
 
 class DealerTest {
